renamed my directories and made a slight change in webpack config.js so that after bundling a bundle.js is created in a dist directory
