+++ index.html
@@ -10,6 +10,6 @@
 		<h1>News Summary App</h1>
 		<h2>Placeholder h2 element</h2>
 
-		<script src='/app.bundle.js/main.js'></script>
+		<script src='dist/bundle.js'></script>
 </body>
 </html>
