+++ src/newsModel.js
