+++ webpack.config.js
@@ -1,9 +1,6 @@
 const path = require("path");
 const dotenv = require('dotenv');
 const webpack = require('webpack'); 
-// import path from 'path'
-// import dotenv from 'dotenv'
-// import webpack from 'webpack'
 
 dotenv.config({path: './.env'})
 
@@ -13,7 +10,8 @@ module.exports = {
 		main: path.resolve(__dirname, 'app.js')
 	},
 	output:{
-		path: path.resolve(__dirname, 'app.bundle.js')
+		path: path.resolve(__dirname, 'dist'),
+		filename: 'bundle.js'
 	},
 	plugins: [
 		new webpack.DefinePlugin({
